ensures valid ethnicity data
